Rename ZookeeperDistributedLock to CuratorZookeeperDistributedLock (#21831)
