+++ mode/type/cluster/repository/provider/zookeeper-curator/src/main/java/org/apache/shardingsphere/mode/repository/cluster/zookeeper/lock/CuratorZookeeperDistributedLock.java
@@ -18,32 +18,40 @@
 package org.apache.shardingsphere.mode.repository.cluster.zookeeper.lock;
 
 import lombok.RequiredArgsConstructor;
-import org.apache.curator.framework.CuratorFramework;
-import org.apache.curator.framework.recipes.locks.InterProcessMutex;
+import org.apache.curator.framework.recipes.locks.InterProcessLock;
 import org.apache.shardingsphere.mode.repository.cluster.lock.DistributedLock;
-import org.apache.shardingsphere.mode.repository.cluster.lock.DistributedLockProvider;
+import org.apache.shardingsphere.mode.repository.cluster.zookeeper.handler.CuratorZookeeperExceptionHandler;
 
-import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.Objects;
+import java.util.concurrent.TimeUnit;
 
 /**
- * Zookeeper distributed lock provider.
+ * Curator ZooKeeper distributed lock.
  */
 @RequiredArgsConstructor
-public final class ZookeeperDistributedLockProvider implements DistributedLockProvider {
+public final class CuratorZookeeperDistributedLock implements DistributedLock {
     
-    private final Map<String, ZookeeperDistributedLock> locks = new LinkedHashMap<>();
-    
-    private final CuratorFramework client;
+    private final InterProcessLock lock;
     
     @Override
-    public synchronized DistributedLock getDistributedLock(final String lockKey) {
-        ZookeeperDistributedLock result = locks.get(lockKey);
-        if (Objects.isNull(result)) {
-            result = new ZookeeperDistributedLock(new InterProcessMutex(client, lockKey));
-            locks.put(lockKey, result);
+    public boolean tryLock(final long timeout) {
+        try {
+            return lock.acquire(timeout, TimeUnit.MILLISECONDS);
+            // CHECKSTYLE:OFF
+        } catch (final Exception ex) {
+            // CHECKSTYLE:ON
+            CuratorZookeeperExceptionHandler.handleException(ex);
         }
-        return result;
+        return false;
+    }
+    
+    @Override
+    public void unlock() {
+        try {
+            lock.release();
+            // CHECKSTYLE:OFF
+        } catch (final Exception ex) {
+            // CHECKSTYLE:ON
+            CuratorZookeeperExceptionHandler.handleException(ex);
+        }
     }
 }
